mui1 removal except where it's still required.
fake recipe maps will need updating once they stop only asking the mui1 for overlays.
