+++ src/main/java/com/gtnewhorizon/cropsnh/compatibility/NEI/CropsNHNEIHandler.java
@@ -3,6 +3,7 @@ package com.gtnewhorizon.cropsnh.compatibility.NEI;
 import java.util.HashMap;
 
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.ResourceLocation;
 
 import org.lwjgl.opengl.GL11;
 
@@ -14,14 +15,14 @@ import codechicken.nei.recipe.HandlerInfo;
 import codechicken.nei.recipe.TemplateRecipeHandler;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import gregtech.api.enums.Mods;
 import gregtech.api.gui.GUIColorOverride;
-import gregtech.api.gui.modularui.GTUITextures;
 
 @SideOnly(Side.CLIENT)
 public abstract class CropsNHNEIHandler extends TemplateRecipeHandler {
 
     protected static final GUIColorOverride colorOverride = GUIColorOverride
-        .get(GTUITextures.BACKGROUND_NEI_SINGLE_RECIPE.location);
+        .get(new ResourceLocation(Mods.GregTech.ID, "gui/background/nei_single_recipe.png"));
 
     private static final HashMap<Class<? extends CropsNHNEIHandler>, Boolean> handlerStatuses = new HashMap<>();
